Implement state and measurements
