+++ src/qibo/gates/__init__.py
@@ -1,3 +1,4 @@
 from qibo.gates.gates import *
+from qibo.gates.measurements import *
 from qibo.gates.special import *
 from qibo.gates.channels import *
